feat(file-upload): improve file upload
