+++ packages/frameworks/react/src/file-upload/file-upload.stories.tsx
@@ -20,7 +20,7 @@ export const Basic = () => (
       {(files) =>
         files.map((file, id) => (
           <FileUpload.Item key={id} file={file}>
-            <FileUpload.ItemPreview />
+            <FileUpload.ItemPreviewImage />
             <FileUpload.ItemName />
             <FileUpload.ItemSizeText />
             <FileUpload.ItemDeleteTrigger>X</FileUpload.ItemDeleteTrigger>
