+++ packages/frameworks/solid/src/file-upload/file-upload-item-preview-image.tsx
@@ -4,9 +4,9 @@ import { ark, type HTMLArkProps } from '../factory'
 import { useFileUploadContext } from './file-upload-context'
 import { useFileUploadItemContext } from './file-upload-item-context'
 
-export interface FileUploadItemPreviewProps extends HTMLArkProps<'img'> {}
+export interface FileUploadItemPreviewImageProps extends HTMLArkProps<'img'> {}
 
-export const FileUploadItemPreview = (props: FileUploadItemPreviewProps) => {
+export const FileUploadItemPreviewImage = (props: FileUploadItemPreviewImageProps) => {
   const api = useFileUploadContext()
   const item = useFileUploadItemContext()
   const [url, setUrl] = createSignal<string>('')
