+++ packages/frameworks/vue/src/file-upload/file-upload.story.vue
@@ -13,7 +13,7 @@ import './file-upload.css'
         <FileUpload.Trigger>Choose file(s)</FileUpload.Trigger>
         <FileUpload.ItemGroup v-slot="files">
           <FileUpload.Item v-for="file in files" :file="file" :key="file">
-            <FileUpload.ItemPreview />
+            <FileUpload.ItemPreviewImage />
             <FileUpload.ItemName />
             <FileUpload.ItemSizeText />
             <FileUpload.ItemDeleteTrigger>X</FileUpload.ItemDeleteTrigger>
