+++ packages/website/src/components/ui/file-upload.tsx
@@ -18,7 +18,11 @@ const FileUploadItemDeleteTrigger = withContext(
 )
 const FileUploadItemGroup = withContext(styled(Ark.FileUpload.ItemGroup), 'itemGroup')
 const FileUploadItemName = withContext(styled(Ark.FileUpload.ItemName), 'itemName')
-const FileUploadItemPreview = withContext(styled(Ark.FileUpload.ItemPreview), 'itemPreview')
+// TODO rename part
+const FileUploadItemPreviewImage = withContext(
+  styled(Ark.FileUpload.ItemPreviewImage),
+  'itemPreview',
+)
 const FileUploadItemSizeText = withContext(styled(Ark.FileUpload.ItemSizeText), 'itemSizeText')
 const FileUploadLabel = withContext(styled(Ark.FileUpload.Label), 'label')
 
@@ -29,7 +33,7 @@ export const FileUpload = Object.assign(FileUploadRoot, {
   ItemDeleteTrigger: FileUploadItemDeleteTrigger,
   ItemGroup: FileUploadItemGroup,
   ItemName: FileUploadItemName,
-  ItemPreview: FileUploadItemPreview,
+  ItemPreviewImage: FileUploadItemPreviewImage,
   ItemSizeText: FileUploadItemSizeText,
   Label: FileUploadLabel,
   Trigger: FileUplaodTrigger,
